+++ frontend/tests/acceptance/helpers/globals.js
